Split examples into two notebooks
